api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
psycopg[binary]==3.2.3
paho-mqtt==2.1.0
pycryptodome==3.23.0
